chore: Remove old Sell flow part 2 (#14407)
