--- src/Apps/Consign/Routes/SubmissionFlow/Utils/types.ts
@@ -1,5 +1,0 @@
-export interface UtmParams {
-  utmTerm?: string
-  utmMedium?: string
-  utmSource?: string
-}
