+++ src/Apps/MyCollection/Routes/PriceEstimate/PriceEstimateContactInformation.tsx
@@ -1,6 +1,17 @@
 import { ActionType } from "@artsy/cohesion"
+import ArtsyLogoIcon from "@artsy/icons/ArtsyLogoIcon"
 import { Button, Flex, Spacer, Text, useToasts } from "@artsy/palette"
+import { PriceEstimateContactInformation_artwork$data } from "__generated__/PriceEstimateContactInformation_artwork.graphql"
+import { PriceEstimateContactInformation_me$data } from "__generated__/PriceEstimateContactInformation_me.graphql"
 import { AppContainer } from "Apps/Components/AppContainer"
+import {
+  contactInformationValidationSchema,
+  validate,
+} from "Apps/Consign/Routes/SubmissionFlow/Utils/validation"
+import {
+  ContactInformationFormFragmentContainer,
+  ContactInformationFormModel,
+} from "Apps/MyCollection/ContactInformation/ContactInformationForm"
 import { useRequestPriceEstimate } from "Apps/MyCollection/Routes/PriceEstimate/Mutations/useRequestPriceEstimate"
 import { BackLink } from "Components/Links/BackLink"
 import { MetaTags } from "Components/MetaTags"
@@ -10,17 +21,6 @@ import { useTracking } from "react-tracking"
 import { RouterLink } from "System/Components/RouterLink"
 import { useRouter } from "System/Hooks/useRouter"
 import { COUNTRY_CODES } from "Utils/countries"
-import { PriceEstimateContactInformation_artwork$data } from "__generated__/PriceEstimateContactInformation_artwork.graphql"
-import { PriceEstimateContactInformation_me$data } from "__generated__/PriceEstimateContactInformation_me.graphql"
-import {
-  contactInformationValidationSchema,
-  validate,
-} from "Apps/Consign/Routes/SubmissionFlow/Utils/validation"
-import ArtsyLogoIcon from "@artsy/icons/ArtsyLogoIcon"
-import {
-  ContactInformationFormFragmentContainer,
-  ContactInformationFormModel,
-} from "Apps/MyCollection/ContactInformation/ContactInformationForm"
 
 const getContactInformationFormInitialValues = (
   me: PriceEstimateContactInformation_me$data
